Moved RunServer to edu.umass.cs.gnsserver.utils/RunCommand. Also removed some warnings.
