+++ src/edu/umass/cs/gnsserver/utils/RunCommand.java
@@ -1,7 +1,23 @@
-/* Copyright (C) 2015 University of Massachusetts All Rights Reserved
- * 
- * Initial developer(s): Westy. */
-package edu.umass.cs.gnsclient.client.integrationtests;
+/*
+ *
+ *  Copyright (c) 2015 University of Massachusetts
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License"); you
+ *  may not use this file except in compliance with the License. You
+ *  may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ *  implied. See the License for the specific language governing
+ *  permissions and limitations under the License.
+ *
+ *  Initial developer(s): Westy
+ *
+ */
+package edu.umass.cs.gnsserver.utils;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -9,7 +25,6 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.Timer;
-import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
 
 /**
@@ -17,7 +32,7 @@ import java.util.concurrent.TimeUnit;
  * 
  * @author westy, arun
  */
-public class RunServer {
+public class RunCommand {
 
 	private static final Timer timer = new Timer(true);
 	private static final int MAX_LINES = 1024;
